move simulation logic to its own class
